hotfixed andy's driver code

+++ source/gatery/scl/sim/SimMemoryMap.cpp
@@ -62,7 +62,7 @@ uint32_t SimulationFiberMapped32BitTileLink::readU32(size_t addr) const
 		auto [val, def, err] = co_await m_linkModel.get(addr, 2, m_clock);
 		if (err) throw std::runtime_error("Bus error!");
 		if (!def) throw std::runtime_error("Undefined value!");
-		co_return val;
+		co_return (uint32_t) val;
 	});
 }
